+++ php_fpm/datadog_checks/php_fpm/php_fpm.py
@@ -1,14 +1,41 @@
 # (C) Datadog, Inc. 2018
 # All rights reserved
 # Licensed under Simplified BSD License (see LICENSE)
+import json
 import random
 import time
 
 import requests
+from flup.client.fcgi_app import FCGIApp
+from six import PY3, StringIO, iteritems, string_types
+from six.moves.urllib.parse import urlparse
 
 from datadog_checks.checks import AgentCheck
+from datadog_checks.config import is_affirmative
 from datadog_checks.utils.headers import headers
-from datadog_checks.config import _is_affirmative
+
+# Relax param filtering
+FCGIApp._environPrefixes.extend(('DOCUMENT_', 'SCRIPT_'))
+
+# Flup as of 1.0.3 is not fully compatible with Python 3 yet.
+# This fixes that for our use case.
+# https://hg.saddi.com/flup-py3.0/file/tip/flup/client/fcgi_app.py
+if PY3:
+    import socket
+
+    def get_connection(self):
+        if self._connect is not None:
+            if isinstance(self._connect, string_types):
+                sock = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
+                sock.connect(self._connect)
+            elif hasattr(socket, 'create_connection'):
+                sock = socket.create_connection(self._connect)
+            else:
+                sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+                sock.connect(self._connect)
+            return sock
+    FCGIApp._getConnection = get_connection
+
 
 DEFAULT_TIMEOUT = 20
 
@@ -42,6 +69,7 @@ class PHPFPMCheck(AgentCheck):
     def check(self, instance):
         status_url = instance.get('status_url')
         ping_url = instance.get('ping_url')
+        use_fastcgi = is_affirmative(instance.get('use_fastcgi', False))
         ping_reply = instance.get('ping_reply')
 
         auth = None
@@ -53,7 +81,7 @@ class PHPFPMCheck(AgentCheck):
 
         timeout = instance.get('timeout', DEFAULT_TIMEOUT)
 
-        disable_ssl_validation = _is_affirmative(instance.get('disable_ssl_validation', False))
+        disable_ssl_validation = is_affirmative(instance.get('disable_ssl_validation', False))
 
         if user and password:
             auth = (user, password)
@@ -64,37 +92,44 @@ class PHPFPMCheck(AgentCheck):
         pool = None
         if status_url is not None:
             try:
-                pool = self._process_status(status_url, auth, tags, http_host, timeout, disable_ssl_validation)
+                pool = self._process_status(
+                    status_url, auth, tags, http_host, timeout, disable_ssl_validation, use_fastcgi
+                )
             except Exception as e:
                 self.log.error("Error running php_fpm check: {}".format(e))
 
         if ping_url is not None:
-            self._process_ping(ping_url, ping_reply, auth, tags, pool, http_host, timeout, disable_ssl_validation)
+            self._process_ping(
+                ping_url, ping_reply, auth, tags, pool, http_host, timeout, disable_ssl_validation, use_fastcgi
+            )
 
-    def _process_status(self, status_url, auth, tags, http_host, timeout, disable_ssl_validation):
+    def _process_status(self, status_url, auth, tags, http_host, timeout, disable_ssl_validation, use_fastcgi):
         data = {}
         try:
-            # TODO: adding the 'full' parameter gets you per-process detailed
-            # informations, which could be nice to parse and output as metrics
-            max_attempts = 3
-            for i in range(max_attempts):
-                resp = requests.get(status_url, auth=auth, timeout=timeout,
-                                    headers=headers(self.agentConfig, http_host=http_host),
-                                    verify=not disable_ssl_validation, params={'json': True})
+            if use_fastcgi:
+                data = json.loads(self.request_fastcgi(status_url))
+            else:
+                # TODO: adding the 'full' parameter gets you per-process detailed
+                # informations, which could be nice to parse and output as metrics
+                max_attempts = 3
+                for i in range(max_attempts):
+                    resp = requests.get(status_url, auth=auth, timeout=timeout,
+                                        headers=headers(self.agentConfig, http_host=http_host),
+                                        verify=not disable_ssl_validation, params={'json': True})
 
-                # Exponential backoff, wait at most (max_attempts - 1) times in case we get a 503.
-                # Delay in seconds is (2^i + random amount of seconds between 0 and 1)
-                # 503s originated here: https://github.com/php/php-src/blob/d84ef96/sapi/fpm/fpm/fpm_status.c#L96
-                if resp.status_code == 503 and i < max_attempts - 1:
-                    # retry
-                    time.sleep(2**i + random.random())
-                    continue
+                    # Exponential backoff, wait at most (max_attempts - 1) times in case we get a 503.
+                    # Delay in seconds is (2^i + random amount of seconds between 0 and 1)
+                    # 503s originated here: https://github.com/php/php-src/blob/d84ef96/sapi/fpm/fpm/fpm_status.c#L96
+                    if resp.status_code == 503 and i < max_attempts - 1:
+                        # retry
+                        time.sleep(2**i + random.random())
+                        continue
 
-                resp.raise_for_status()
-                data = resp.json()
+                    resp.raise_for_status()
+                    data = resp.json()
 
-                # successfully got a response, exit the backoff system
-                break
+                    # successfully got a response, exit the backoff system
+                    break
         except Exception as e:
             self.log.error("Failed to get metrics from {}: {}".format(status_url, e))
             raise
@@ -104,13 +139,13 @@ class PHPFPMCheck(AgentCheck):
         if http_host is not None:
             metric_tags += ["http_host:{0}".format(http_host)]
 
-        for key, mname in self.GAUGES.iteritems():
+        for key, mname in iteritems(self.GAUGES):
             if key not in data:
                 self.log.warn("Gauge metric {0} is missing from FPM status".format(key))
                 continue
             self.gauge(mname, int(data[key]), tags=metric_tags)
 
-        for key, mname in self.MONOTONIC_COUNTS.iteritems():
+        for key, mname in iteritems(self.MONOTONIC_COUNTS):
             if key not in data:
                 self.log.warn("Counter metric {0} is missing from FPM status".format(key))
                 continue
@@ -119,7 +154,9 @@ class PHPFPMCheck(AgentCheck):
         # return pool, to tag the service check with it if we have one
         return pool_name
 
-    def _process_ping(self, ping_url, ping_reply, auth, tags, pool_name, http_host, timeout, disable_ssl_validation):
+    def _process_ping(
+        self, ping_url, ping_reply, auth, tags, pool_name, http_host, timeout, disable_ssl_validation, use_fastcgi
+    ):
         if ping_reply is None:
             ping_reply = 'pong'
 
@@ -129,14 +166,18 @@ class PHPFPMCheck(AgentCheck):
 
         try:
             # TODO: adding the 'full' parameter gets you per-process detailed
-            # informations, which could be nice to parse and output as metrics
-            resp = requests.get(ping_url, auth=auth, timeout=timeout,
-                                headers=headers(self.agentConfig, http_host=http_host),
-                                verify=not disable_ssl_validation)
-            resp.raise_for_status()
+            # information, which could be nice to parse and output as metrics
+            if use_fastcgi:
+                response = self.request_fastcgi(ping_url).decode('utf-8')
+            else:
+                resp = requests.get(ping_url, auth=auth, timeout=timeout,
+                                    headers=headers(self.agentConfig, http_host=http_host),
+                                    verify=not disable_ssl_validation)
+                resp.raise_for_status()
+                response = resp.text
 
-            if ping_reply not in resp.text:
-                raise Exception("Received unexpected reply to ping: {}".format(resp.text))
+            if ping_reply not in response:
+                raise Exception("Received unexpected reply to ping: {}".format(response))
 
         except Exception as e:
             self.log.error("Failed to ping FPM pool {} on URL {}: {}".format(pool_name, ping_url, e))
@@ -144,3 +185,43 @@ class PHPFPMCheck(AgentCheck):
                                message=str(e))
         else:
             self.service_check(self.SERVICE_CHECK_NAME, AgentCheck.OK, tags=sc_tags)
+
+    @classmethod
+    def request_fastcgi(cls, url):
+        parsed_url = urlparse(url)
+
+        hostname = parsed_url.hostname
+        if hostname == 'localhost':
+            hostname = '127.0.0.1'
+
+        port = str(parsed_url.port or 9000)
+        route = parsed_url.path
+
+        query_string = 'json' if route == '/status' else ''
+
+        env = {
+            'CONTENT_LENGTH': '0',
+            'CONTENT_TYPE': '',
+            'DOCUMENT_ROOT': '/',
+            'GATEWAY_INTERFACE': 'FastCGI/1.1',
+            'QUERY_STRING': query_string,
+            'REDIRECT_STATUS': '200',
+            'REMOTE_ADDR': '127.0.0.1',
+            'REMOTE_PORT': '80',
+            'REQUEST_METHOD': 'GET',
+            'REQUEST_URI': route,
+            'SCRIPT_FILENAME': route,
+            'SCRIPT_NAME': route,
+            'SERVER_ADDR': hostname,
+            'SERVER_NAME': hostname,
+            'SERVER_PORT': port,
+            'SERVER_PROTOCOL': 'HTTP/1.1',
+            'SERVER_SOFTWARE': 'datadog-php-fpm',
+            'wsgi.errors': StringIO(),
+            'wsgi.input': StringIO(),
+        }
+
+        fcgi = FCGIApp(host=hostname, port=port)
+
+        # Return first response
+        return fcgi(env, lambda *args, **kwargs: '')[0]