+++ vsphere/datadog_checks/vsphere/vsphere.py
@@ -89,6 +89,10 @@ def atomic_method(method):
     return wrapper
 
 
+class ConnectionError(Exception):
+    pass
+
+
 class VSphereCheck(AgentCheck):
     """ Get performance metrics from a vCenter server and upload them to Datadog
     References:
@@ -224,16 +228,7 @@ class VSphereCheck(AgentCheck):
         now = time.time()
         return now - self.cache_times[i_key][entity][LAST] > self.cache_times[i_key][entity][INTERVAL]
 
-    def _get_server_instance(self, instance):
-        i_key = self._instance_key(instance)
-        tags = instance.get('tags', [])
-
-        service_check_tags = [
-            'vcenter_server:{0}'.format(instance.get('name')),
-            'vcenter_host:{0}'.format(instance.get('host')),
-        ] + tags
-        service_check_tags = list(set(service_check_tags))
-
+    def _smart_connect(self, instance, service_check_tags):
         # Check for ssl configs and generate an appropriate ssl context object
         ssl_verify = instance.get('ssl_verify', True)
         ssl_capath = instance.get('ssl_capath', None)
@@ -252,34 +247,55 @@ class VSphereCheck(AgentCheck):
                            "verification. You cannot do both. Proceeding with "
                            "disabling ssl verification.")
 
-        if i_key not in self.server_instances:
-            try:
-                # Object returned by SmartConnect is a ServiceInstance
-                #   https://code.vmware.com/apis/358/vsphere#/doc/vim.ServiceInstance.html
-                server_instance = connect.SmartConnect(
-                    host=instance.get('host'),
-                    user=instance.get('username'),
-                    pwd=instance.get('password'),
-                    sslContext=context if not ssl_verify or ssl_capath else None
-                )
-            except Exception as e:
-                err_msg = "Connection to %s failed: %s" % (instance.get('host'), e)
-                self.service_check(self.SERVICE_CHECK_NAME, AgentCheck.CRITICAL,
-                                   tags=service_check_tags, message=err_msg)
-                raise Exception(err_msg)
+        try:
+            # Object returned by SmartConnect is a ServerInstance
+            #   https://www.vmware.com/support/developer/vc-sdk/visdk2xpubs/ReferenceGuide/vim.ServiceInstance.html
+            server_instance = connect.SmartConnect(
+                host=instance.get('host'),
+                user=instance.get('username'),
+                pwd=instance.get('password'),
+                sslContext=context if not ssl_verify or ssl_capath else None
+            )
+        except Exception as e:
+            err_msg = "Connection to {} failed: {}".format(instance.get('host'), e)
+            self.service_check(self.SERVICE_CHECK_NAME, AgentCheck.CRITICAL,
+                               tags=service_check_tags, message=err_msg)
+            raise ConnectionError(err_msg)
 
-            self.server_instances[i_key] = server_instance
+        # Check that we have sufficient permission for the calls we need to make
+        try:
+            server_instance.CurrentTime()
+        except Exception as e:
+            err_msg = (
+                "A connection to {} can be established, but performing operations on the server fails: {}"
+            ).format(instance.get('host'), instance.get('username'), e)
+            self.service_check(self.SERVICE_CHECK_NAME, AgentCheck.CRITICAL,
+                               tags=service_check_tags, message=err_msg)
+            raise ConnectionError(err_msg)
+
+        return server_instance
+
+    def _get_server_instance(self, instance):
+        i_key = self._instance_key(instance)
+        tags = instance.get('tags', [])
+
+        service_check_tags = [
+            'vcenter_server:{0}'.format(instance.get('name')),
+            'vcenter_host:{0}'.format(instance.get('host')),
+        ] + tags
+        service_check_tags = list(set(service_check_tags))
+
+        if i_key not in self.server_instances:
+            self.server_instances[i_key] = self._smart_connect(instance, service_check_tags)
 
         # Test if the connection is working
         try:
-            self.server_instances[i_key].RetrieveContent()
+            self.server_instances[i_key].CurrentTime()
             self.service_check(self.SERVICE_CHECK_NAME, AgentCheck.OK,
                                tags=service_check_tags)
-        except Exception as e:
-            err_msg = "Connection to %s died unexpectedly: %s" % (instance.get('host'), e)
-            self.service_check(self.SERVICE_CHECK_NAME, AgentCheck.CRITICAL,
-                               tags=service_check_tags, message=err_msg)
-            raise Exception(err_msg)
+        except Exception:
+            # Try to reconnect. If the connection is definitely broken, this will send CRITICAL service check and raise
+            self.server_instances[i_key] = self._smart_connect(instance, service_check_tags)
 
         return self.server_instances[i_key]